Not initializing container on restart

+++ run.sh
@@ -2,11 +2,15 @@
 
 set -e
 
-envtpl /etc/nginx/nginx.conf.tpl
-envtpl /grafana/config.js.tpl
+if [ ! -e /.initialized ]; then
+    touch /.initialized
 
-if [ ! -z "${GRAFANA_BASIC_AUTH_LOGIN}" ]; then
-    echo "${GRAFANA_BASIC_AUTH_LOGIN}:${GRAFANA_BASIC_AUTH_PASSWORD}" > /etc/nginx/grafana.htpasswd
+    envtpl /etc/nginx/nginx.conf.tpl
+    envtpl /grafana/config.js.tpl
+
+    if [ ! -z "${GRAFANA_BASIC_AUTH_LOGIN}" ]; then
+        echo "${GRAFANA_BASIC_AUTH_LOGIN}:${GRAFANA_BASIC_AUTH_PASSWORD}" > /etc/nginx/grafana.htpasswd
+    fi
 fi
 
 exec nginx
